rename card component, add progress bar click option
