+++ components/admin_console/cloud_import/widgets/progress_bar.scss
@@ -21,3 +21,7 @@
         color: var(--sys-center-channel-color);
     }
 }
+
+.clickable {
+    cursor: pointer;
+}
