+++ components/admin_console/cloud_import/widgets/status_legend.scss
