+++ components/common/card_container.scss
@@ -1,4 +1,4 @@
-.CloudCard {
+.CardContainer {
     background-color: var(--sys-center-channel-bg);
     box-shadow: 0px 2px 3px rgba(0, 0, 0, 0.08);
     border: 1px solid rgba(var(--sys-center-channel-color-rgb), 0.08);
